requirements.txt
pydantic-ai
openai
pinecone-client
python-dotenv 